don't index an array with char. pointed out by gcc

+++ main.c
@@ -1120,7 +1120,8 @@ main(int argc, char **argv, char **envp)
 #endif				/* USE_MOUSE */
 		if (IS_ASCII(c)) {	/* Ascii */
 			if (('0' <= c) && (c <= '9') &&
-			(prec_num || (GlobalKeymap[c] == FUNCNAME_nulcmd))) {
+			(prec_num || (GlobalKeymap[(uint8_t) c] ==
+			    FUNCNAME_nulcmd))) {
 				prec_num = prec_num * 10 + (int) (c - '0');
 				if (prec_num > PREC_LIMIT)
 					prec_num = PREC_LIMIT;
